moves files to other directory
