+++ covid19-dashboard/README.md
@@ -1,15 +1,48 @@
-# dc-covid19-website
-Dashboard to display the top number of COVID-19 regions (State or County) in the USA by cases or deaths.
+# DC COVID-19 Dashboard  
+#### Server APIs:  
+**api/state-data**: returns the data for all States in the US with NYT COVID-19 data.  
+ - downloads in a few ms.  
+  
+**api/county-data**: returns the data for all Counties in the US with NYT COVID-19 data.  
+ - downloads in 1-2 seconds.  
+   
+ **api/country-data**: returns the data for all Countries in the world with WHO COVID-19 data.  
+ - downloads in 1-2 seconds.  
+  
+######  The return types for both APIs are identical, they are of the following type:  
+  
+``{  
+"name": string,  
+'population': int,  
+'containedIn': string,  
+'7DayAverageCases': {iso-date: int},  
+'7DayAverageDeaths': {iso-date: int},  
+'7DayAverageCasesValue': int,  
+'7DayAverageDeathsValue': int,  
+'perCapitaCases': {iso-date: int},  
+'perCapitaDeaths': {iso-date: int},  
+'perCapitaCasesValue': int,  
+'perCapitaDeathsValue': int,  
+'pctChangeCasesValue': int,  
+'pctChangeDeathsValue': int  
+}``  
+  
+  
+#### Example  
 
-To build the project:
-1. Create a config.py file in the covid19-dashboard directory.
-   a. The config file should contain your DataCommons API Key as follows:``API_KEY="YOURAPIKEY"``.
-2. Go to dc-covid19 and execute ``npm run build`` to build the static React files.
-
-To deploy the project to Google Cloud App Engine:
-1. Install Google Cloud Command Line Tools.
-2. Set Up an App Engine Instance.
-3. Login to your App Engine Instance using the Google Cloud Command Line Tools.
-4. Execute ``gcloud app deploy``.
-5. Execute ``gcloud app deploy cron.yaml``.
-
+    {  
+    "name": 'Florida', # region's name.  
+    'population': 21_480_000, # total population.  
+    'containedIn': 'country/USA', # geo_id of the place that directly contains this geoId.  
+    'type': 'State', # type of region: "Country", "State" or "County".  
+    '7DayAverageCases': {'2020-01-01': 10, '2020-01-02': 12}, # time-series data as a dictionary.  
+    '7DayAverageCasesValue': 12, # value for the latest date in the "7DayAverageCases" time-series.  
+    '7DayAverageDeaths': {'2020-01-01': 4, '2020-01-02': 5}, # time-series data as a dictionary.  
+    '7DayAverageDeathsValue': 5, # value for the latest date in the "7DayAverageDeaths" time-series.  
+    'perCapitaCases': {'2020-01-01': 0.46, '2020-01-02': 0.55}, # time-series data as a dictionary.  
+    'perCapitaCasesValue': 0.55, # value for the latest date in the "perCapitaCases" time-series.  
+    'perCapitaDeaths': {'2020-01-01': 0.18, '2020-01-02': 0.23}, # time-series data as a dictionary.  
+    'perCapitaDeathsValue': 0.23, # value for the latest date in the "perCapitaDeaths" time-series.  
+    'pctChangeCasesValue': 19.57, # value representing an increase of cases from date X to latest date.  
+    'pctChangeDeathsValue': 27.78 # value representing an increase of deaths from date X to latest date.  
+    }
